\name{getPAA}
\alias{getPAA}
\title{getPAA}
\description{Piecewise Aggregated approximation of the signal  }
\usage{getPAA(ts, wl = 16)}
\arguments{
  \item{ts}{time series to be converted}
  \item{wl}{number of symbols. After conversion time series will be wl in length.}
}

\value{vector of doubles wl sy}

\author{Anatoly Sorokin}





