\name{hashSAX}
\alias{hashSAX}
\title{#<< signal }

\usage{hashSAX(cp, wl = 16, win = length(ts), verbose = FALSE)}
\arguments{
  \item{cp}{signal}
  \item{wl}{desired length of the string representation}
  \item{win}{sliding window length. Signal will be represented as set of length(ts)-win+1 strings of wl characters each.}
  \item{verbose}{if TRUE print progress indicator}
}



\author{Anatoly Sorokin}





