\name{hSAX2signal}
\alias{hSAX2signal}
\title{hSAX2signal}
\description{function return double representation of the hSAX signal of the same width as original signal}
\usage{hSAX2signal(str, win = 48)}
\arguments{
  \item{str}{hSAX string}
  \item{win}{width of the original signal}
}



\author{Anatoly Sorokin}





