\name{hSAXbase}
\alias{hSAXbase}
\title{hSAXbase}
\description{define basic properties of hexSAX string:
size of the alphabet, alphabet itself  and break points for convertison}
\usage{hSAXbase(alphasize = 16)}
\arguments{
  \item{alphasize}{size of the alphabet, should be less then 16}
}



\author{Anatoly Sorokin}





