\name{hsaxDist}
\alias{hsaxDist}
\title{hsaxDist}
\description{distance between two hSAX strings  }
\usage{hsaxDist(x, y, alphasize = 16, md)}
\arguments{
  \item{x}{first string}
  \item{y}{second string}
  \item{alphasize}{expected alphabet size}
  \item{md}{dissimilarity matrix for SAX alphabet}
}



\author{Anatoly Sorokin}





