\name{iSAXbase}
\alias{iSAXbase}
\title{iSAXbase}
\description{define basic properties of iSAX string:
the cardinality of the alphabet, size of the alphabet, 
alphabet itself  and break points for convertison}
\usage{iSAXbase(card = 8)}
\arguments{
  \item{card}{cardinality of the alphabet, should be less then 8}
}



\author{Anatoly Sorokin}





