\name{iSAXr-package}
\alias{iSAXr-package}
\docType{package}
\title{What the package does (short line)}
\description{More about what it does (maybe more than one line)}
\details{
\tabular{ll}{Package: \tab iSAXr\cr
Type: \tab Package\cr
Title: \tab What the package does (short line)\cr
Version: \tab 0.0-3\cr
Date: \tab 2013-02-05\cr
Author: \tab Anatoly Sorokin\cr
Maintainer: \tab Anatoly Sorokin <lptolik@gmail.com>\cr
License: \tab GPL(>=2)\cr
Depends: \tab testthat\cr
Suggests: \tab \cr}
}
\author{Anatoly Sorokin}

\keyword{ package }


