\name{mov.train}
\alias{mov.train}
\docType{data}
\title{The Internet Movie Database Train Set}
\description{
Training set of the Internet Movie Database IMDb from \url{http://www.imdb.com} manually tagged. The original text files can be downloaded from \url{http://ai.stanford.edu/~amaas/data/sentiment}.
}
\usage{data(mov.train)}
\references{
Maas, Daly,  Pham, Huang, Ng, Potts (2011) Learning Word Vectors for Sentiment Analysis. The 49th Annual Meeting of the Association for Computational Linguistics (ACL 2011).
}
%\examples{
%
%}
\keyword{datasets}
