\name{hsaxDist}
\alias{hsaxDist}
\title{hsaxDist}
\description{distance between two hSAX strings  }
\usage{hsaxDist(x, y, base = hSAXbase(), md)}
\arguments{
  \item{x}{first string}
  \item{y}{second string}
  \item{base}{SAXbase object}
  \item{md}{dissimilarity matrix for SAX alphabet}
}



\author{Anatoly Sorokin}





