\name{iSAX-package}
\alias{iSAX-package}
\docType{package}
\title{
\packageTitle{iSAX}
}
\description{
\packageDescription{iSAX}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{iSAX}
\packageIndices{iSAX}
}
\author{
\packageAuthor{iSAX}
Maintainer: \packageMaintainer{iSAX}
}
\references{
Iacus, S.M., Curini, L., Ceron, A. (2015) iSA (U.S. provisional patent application No. 62/215264)
Ceron, A., Curini, L., Iacus, S.M. (2016) iSA: A fast, scalable and accurate algorithm for sentiment analysis of social media content, Information Sciences, V. 367-368, p. 105-124.

}
\keyword{ package }
\seealso{
\code{\link{iSA}}
\code{\link{prep.data}}
}
